Added telent test module

+++ src/dev/mem/mem.go
@@ -1,5 +1,5 @@
 // Package Mem,
-// Not much to do read/write bytes
+// read/write bytes, halfwords, workds, and boublewords
 
 package mem
 
@@ -24,20 +24,19 @@ func (m *Mem) SetEndian(o binary.ByteOrder) {
 	m.endian = o
 }
 
-func (m *Mem) Read8At(off uint32) (uint8, error) {
+func (m *Mem) Read8At(off uint64) (uint8, error) {
 	return m.buf[off], nil
 }
 
-func (m *Mem) Read16At(off uint32) (uint16, error) {
-	//var v uint32
+func (m *Mem) Read16At(off uint64) (uint16, error) {
 	return m.endian.Uint16(m.buf[off:]), nil
 }
 
-func (m *Mem) Read32At(off uint32) (uint32, error) {
+func (m *Mem) Read32At(off uint64) (uint32, error) {
 	return m.endian.Uint32(m.buf[off:]), nil
 }
 
-func (m *Mem) Read64At(off uint32) (uint64, error) {
+func (m *Mem) Read64At(off uint64) (uint64, error) {
 	return m.endian.Uint64(m.buf[off:]), nil
 }
 
@@ -69,6 +68,8 @@ func (m *Mem) ReadAt(p []byte, off uint64) (n int, e error) {
 	return
 }
 
+// The bus will call this function, so does CPU,
+// if
 func (m *Mem) WriteAt(p []byte, off uint64) (n int, e error) {
 	n = copy(m.buf[off:], p)
 	if n != len(p) {
